app: Add logic to handle $ref in parameters BNCH-125141 BNCH-125943 (#231)
In December of 2020, when we diverged from the upstream branch, the client generator could not yet handle `$ref` in the parameters list. They added that feature in [August 2022](https://github.com/openapi-generators/openapi-python-client/commit/95b7eb33b69485811bdf2ca2aafdde6d86b01298) but we never pulled it in.

In this PR, I try to pull in the same implementation from the upstream while modifying our repo as little as possible. Only the tests are really hand-written.

* Minimal changes to enable debugging

* Improve configuration for vscode development

- Set ruler to 120 characters, which is the configured value in black
- Set up vscode to use pytest with the same configuration as task unit
- Turn off mypy in tests. They are so noncompliant that it's useless

* Copy ParameterError from upstream

* Copy ClassName from upstream

* Copy parameter reference handling from upstream

This builds a library of parameter references from the openapi spec and
gives a utility for looking them up.

* Copy build_parameters from upstream

This builds the actual Parameters object from the OpenAPI spec, using
the logic we just added to schemas.

* Copy parameter reference logic from upstream

Intead of ignoring reference-typed parameters, look them up in the
Parameters object.

* Update unit tests

* Update version to 1.0.7

* Remove explicit local python interpreter

* Format launch.json